Add VIPER style context bar
Uses the same config.create_title_from_context_style as
the title bar does. If set to 'openaire' context is formatted as desired.
Otherwise it behaves as before

Added placeholder elements to headstart template for data to go it.

Added english localized labels for human text.

Notable deviation from mock up is lack of funding stream. This is not
returned from the OA projects API. If needed we'll have to get this elsewhere.