Small refactor of filtering
Improved some function names and corrected and added some new
comments to aid understanding.

Remove unreachable code. paper_abstract is now always defined because
we set to a default in io.js.

Unreacable code was never run but is misleading to the reader since it
presumabling was written at a time when filtering was an additive action
not a subtractive one.