Force single instance of Headstart
Fixes dropdown in local_files.

Probably fixes scaling in Viper.

Prior to the webpack 4 version increase it seems that a single
instance of headstart was used throughout the application.

This seems to have been been possible by manipulating the
headstart alias imported in app.js. It wasn't possible to continue
this pattern with webpack 4. I'm unsure the exact reason why but
doing this is probably not advised anyway.

This patch forces an explict global headstart object and ensures
the callers in mediator, canvas and scale all use it.

This is probably not the ideal solution. It might be better to
instead properly inject headstart into the places that need it.