Setup some linting for components
Linting is currently restricted to only modern code
because there are many violations of these linting rules
in the upstream code

The PR introduces the same style of linting
that is used by create-react-app. It appears
to be the industry standard way to lint these kind of apps.

Updated developer documentation to make how to lint clear.